customnerd-backend/requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.13.3
bio==1.6.2
biopython==1.85
biothings-client==0.3.1
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
effdet==0.4.1
emoji==2.11.0
et-xmlfile==1.1.0
eutils==0.6.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fake-useragent==1.5.1
fastapi==0.115.10
filelock==3.13.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
google-genai
googleapis-common-protos==1.63.0
gprofiler-official==1.0.0
greenlet==3.0.3
# grpcio==1.62.1
h11==0.14.0
habanero==1.2.6
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
json5
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.48
layoutparser==0.3.4
Levenshtein==0.25.1
lxml==5.2.1
lxml_html_clean==0.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
metapub==0.5.5
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
mygene==3.2.2
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
networkx==3.2.1
nltk==3.8.1
numpy==2.2.3
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.20.1
openai==1.14.1
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson 
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.14.0
pillow
pillow_heif
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
portalocker==2.8.2
posthog==3.5.0
protobuf==4.25.3
pulsar-client==3.6.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.0
pydantic_core
Pygments==2.17.2
pymupdf==1.25.3
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.29.0
PyPika==0.48.9
pyproject-api==1.6.1
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pytesseract==0.3.10
pytest==8.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-Levenshtein==0.25.1
python-magic==0.4.27
python-multipart==0.0.20
python-pptx==0.6.23
pytz
PyYAML==6.0.1
rapidfuzz==3.8.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.29
sse-starlette==2.2.1
starlette==0.46.0
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==0.9.16
tokenizers==0.15.2
tomli
# torch #Not working
# torchvision==0.17.2  #Not working
tox==4.14.2
tqdm==4.66.2
transformers==4.39.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2025.1
Unidecode==1.3.8
unstructured
unstructured-client==0.18.0
# unstructured-inference==0.7.25  #Not working    
unstructured.pytesseract==0.3.12
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
watchfiles
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.18.1
